allow overriding how certain configuration variables get displayed to the user
Some configuration variables are not part of the build matrix,
and some other need a proper name instead of a raw value. This
commit allows to set a 'metadata' flag on values to give them
name. Setting it to nil will make this particular value ignored
w.r.t. image name and display